requirements.txt
future==0.17.1
neat-python==0.92
numpy==1.17.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pygame==1.9.6
pyglet==1.3.2
PyGObject==3.30.4
pylint==2.4.4
PySimpleSOAP==1.16.2
scipy==1.3.1
